+++ CHANGES.md
@@ -1,3 +1,6 @@
 
+- Using `Option<T>`
+- Using `Catch-all` Patterns and `the _ Placeholder`
+
 ## 0.1.0
 - Initial commit